Refactoring Native type
* Many types depend on `Native`, so initialization uses a shorthand to
ensure it is called first — `shoes_0_native_type_init()`.
* Shoes3/shoes3#333
