+++ shoes/ruby.c
@@ -12,7 +12,7 @@
 #include "shoes/types/types.h"
 #include <math.h>
 
-VALUE cShoes, cApp, cDialog, cTypes, cShoesWindow, cMouse, cCanvas, cFlow, cStack, cMask, cWidget, cImage, cPattern, cBorder, cBackground, cTextBlock, cPara, cBanner, cTitle, cSubtitle, cTagline, cCaption, cInscription, cTextClass, cNative, cProgress, cColor, cResponse, ssNestSlot;
+VALUE cShoes, cApp, cDialog, cTypes, cShoesWindow, cMouse, cCanvas, cFlow, cStack, cMask, cWidget, cImage, cPattern, cBorder, cBackground, cTextBlock, cPara, cBanner, cTitle, cSubtitle, cTagline, cCaption, cInscription, cTextClass, cProgress, cColor, cResponse, ssNestSlot;
 VALUE eImageError, eInvMode, eNotImpl;
 VALUE reHEX_SOURCE, reHEX3_SOURCE, reRGB_SOURCE, reRGBA_SOURCE, reGRAY_SOURCE, reGRAYA_SOURCE, reLF;
 VALUE symAltQuest, symAltSlash, symAltDot, symAltEqual, symAltSemiColon;
@@ -625,14 +625,6 @@ void shoes_cairo_rect(cairo_t *cr, double x, double y, double w, double h, doubl
     cairo_rel_line_to(cr, 0, -h + 2 * r);
     if (r != 0.) cairo_rel_curve_to(cr, 0.0, -rc, r - rc, -r, r, -r);
     cairo_close_path(cr);
-}
-
-void shoes_control_hide_ref(SHOES_CONTROL_REF ref) {
-    if (ref != NULL) shoes_native_control_hide(ref);
-}
-
-void shoes_control_show_ref(SHOES_CONTROL_REF ref) {
-    if (ref != NULL) shoes_native_control_show(ref);
 }
 
 //
@@ -1764,173 +1756,8 @@ VALUE shoes_textblock_string(VALUE self) {
     return rb_str_new(self_t->text->str, self_t->text->len);
 }
 
-//
-// Shoes::Button
-//
-void shoes_control_mark(shoes_control *control) {
-    rb_gc_mark_maybe(control->parent);
-    rb_gc_mark_maybe(control->attr);
-}
-
-static void shoes_control_free(shoes_control *control) {
-    if (control->ref != NULL) shoes_native_control_free(control->ref);
-    RUBY_CRITICAL(free(control));
-}
-
-VALUE shoes_control_new(VALUE klass, VALUE attr, VALUE parent) {
-    shoes_control *control;
-    VALUE obj = shoes_control_alloc(klass);
-    Data_Get_Struct(obj, shoes_control, control);
-    control->attr = attr;
-    control->parent = parent;
-    return obj;
-}
-
-VALUE shoes_control_alloc(VALUE klass) {
-    VALUE obj;
-    shoes_control *control = SHOE_ALLOC(shoes_control);
-    SHOE_MEMZERO(control, shoes_control, 1);
-    obj = Data_Wrap_Struct(klass, shoes_control_mark, shoes_control_free, control);
-    control->attr = Qnil;
-    control->parent = Qnil;
-    return obj;
-}
-
-VALUE shoes_control_focus(VALUE self) {
-    GET_STRUCT(control, self_t);
-//  ATTRSET(self_t->attr, hidden, Qtrue);
-    if (self_t->ref != NULL) shoes_native_control_focus(self_t->ref);
-    return self;
-}
-
-VALUE shoes_control_get_state(VALUE self) {
-    GET_STRUCT(control, self_t);
-    return ATTR(self_t->attr, state);
-}
-
-static VALUE shoes_control_try_state(shoes_control *self_t, VALUE state) {
-    unsigned char cstate;
-    if (NIL_P(state))
-        cstate = CONTROL_NORMAL;
-    else if (TYPE(state) == T_STRING) {
-        if (strncmp(RSTRING_PTR(state), "disabled", 8) == 0)
-            cstate = CONTROL_DISABLED;
-        else if (strncmp(RSTRING_PTR(state), "readonly", 8) == 0)
-            cstate = CONTROL_READONLY;
-        else {
-            shoes_error("control can't have :state of %s\n", RSTRING_PTR(state));
-            return Qfalse;
-        }
-    } else return Qfalse;
-
-    if (self_t->ref != NULL) {
-        if (cstate == CONTROL_NORMAL)
-            shoes_native_control_state(self_t->ref, TRUE, TRUE);
-        else if (cstate == CONTROL_DISABLED)
-            shoes_native_control_state(self_t->ref, FALSE, TRUE);
-        else if (cstate == CONTROL_READONLY)
-            shoes_native_control_state(self_t->ref, TRUE, FALSE);
-    }
-    return Qtrue;
-}
-
-VALUE shoes_control_set_state(VALUE self, VALUE state) {
-    GET_STRUCT(control, self_t);
-    if (shoes_control_try_state(self_t, state))
-        ATTRSET(self_t->attr, state, state);
-    return self;
-}
-
-VALUE shoes_control_temporary_hide(VALUE self) {
-    GET_STRUCT(control, self_t);
-    if (self_t->ref != NULL) shoes_control_hide_ref(self_t->ref);
-    return self;
-}
-
-VALUE shoes_control_hide(VALUE self) {
-    GET_STRUCT(control, self_t);
-    ATTRSET(self_t->attr, hidden, Qtrue);
-    if (self_t->ref != NULL) shoes_control_hide_ref(self_t->ref);
-    return self;
-}
-
-VALUE shoes_control_temporary_show(VALUE self) {
-    GET_STRUCT(control, self_t);
-    if (self_t->ref != NULL) shoes_control_show_ref(self_t->ref);
-    return self;
-}
-
-VALUE shoes_control_show(VALUE self) {
-    GET_STRUCT(control, self_t);
-    ATTRSET(self_t->attr, hidden, Qfalse);
-    if (self_t->ref != NULL) shoes_control_show_ref(self_t->ref);
-    return self;
-}
-
-VALUE shoes_control_remove(VALUE self) {
-    shoes_canvas *canvas;
-    GET_STRUCT(control, self_t);
-    shoes_canvas_remove_item(self_t->parent, self, 1, 0);
-
-    Data_Get_Struct(self_t->parent, shoes_canvas, canvas);
-    if (self_t->ref != NULL) {
-        SHOES_CONTROL_REF ref = self_t->ref;
-        self_t->ref = NULL;
-        shoes_native_control_remove(ref, canvas);
-    }
-    return self;
-}
-
-void shoes_control_check_styles(shoes_control *self_t) {
-    VALUE x = ATTR(self_t->attr, state);
-    shoes_control_try_state(self_t, x);
-}
-
-void shoes_control_send(VALUE self, ID event) {
-    VALUE click;
-    GET_STRUCT(control, self_t);
-
-    if (rb_respond_to(self, s_checked))
-        ATTRSET(self_t->attr, checked, rb_funcall(self, s_checked, 0));
-
-    if (!NIL_P(self_t->attr)) {
-        click = rb_hash_aref(self_t->attr, ID2SYM(event));
-        if (!NIL_P(click))
-            shoes_safe_block(self_t->parent, click, rb_ary_new3(1, self));
-    }
-}
-
-VALUE shoes_control_get_tooltip(VALUE self) {
-    GET_STRUCT(control, self_t);
-    return shoes_native_control_get_tooltip(self_t->ref);
-}
-
-VALUE shoes_control_set_tooltip(VALUE self, VALUE tooltip) {
-    GET_STRUCT(control, self_t);
-    if (self_t->ref != NULL)
-        shoes_native_control_set_tooltip(self_t->ref, tooltip);
-
-    return self;
-}
-
-VALUE shoes_button_group(VALUE self) {
-    GET_STRUCT(control, self_t);
-    if (!NIL_P(self_t->parent)) {
-        shoes_canvas *canvas;
-        VALUE group = ATTR(self_t->attr, group);
-        if (NIL_P(group)) group = self_t->parent;
-        Data_Get_Struct(self_t->parent, shoes_canvas, canvas);
-        return shoes_hash_get(canvas->app->groups, group);
-    }
-    return Qnil;
-}
-
 PLACE_COMMON(canvas)
 TRANS_COMMON(canvas, 0);
-PLACE_COMMON(control)
-CLASS_COMMON(control)
-EVENT_COMMON(control, control, click)
-EVENT_COMMON(control, control, change)
 PLACE_COMMON(image)
 CLASS_COMMON2(image)
 TRANS_COMMON(image, 1);
@@ -2269,24 +2096,6 @@ void shoes_ruby_init() {
     cTagline = rb_define_class_under(cTypes, "Tagline", cTextBlock);
     cCaption = rb_define_class_under(cTypes, "Caption", cTextBlock);
     cInscription = rb_define_class_under(cTypes, "Inscription", cTextBlock);
-
-    cNative  = rb_define_class_under(cTypes, "Native", rb_cObject);
-    rb_define_alloc_func(cNative, shoes_control_alloc);
-    rb_define_method(cNative, "app", CASTHOOK(shoes_canvas_get_app), 0);
-    rb_define_method(cNative, "parent", CASTHOOK(shoes_control_get_parent), 0);
-    rb_define_method(cNative, "style", CASTHOOK(shoes_control_style), -1);
-    rb_define_method(cNative, "displace", CASTHOOK(shoes_control_displace), 2);
-    rb_define_method(cNative, "focus", CASTHOOK(shoes_control_focus), 0);
-    rb_define_method(cNative, "hide", CASTHOOK(shoes_control_hide), 0);
-    rb_define_method(cNative, "show", CASTHOOK(shoes_control_show), 0);
-    rb_define_method(cNative, "state=", CASTHOOK(shoes_control_set_state), 1);
-    rb_define_method(cNative, "state", CASTHOOK(shoes_control_get_state), 0);
-    rb_define_method(cNative, "move", CASTHOOK(shoes_control_move), 2);
-    rb_define_method(cNative, "top", CASTHOOK(shoes_control_get_top), 0);
-    rb_define_method(cNative, "left", CASTHOOK(shoes_control_get_left), 0);
-    rb_define_method(cNative, "width", CASTHOOK(shoes_control_get_width), 0);
-    rb_define_method(cNative, "height", CASTHOOK(shoes_control_get_height), 0);
-    rb_define_method(cNative, "remove", CASTHOOK(shoes_control_remove), 0);
 
     rb_define_method(cApp, "method_missing", CASTHOOK(shoes_app_method_missing), -1);
