+++ shoes/types/text_link.h
@@ -10,8 +10,8 @@
    because link is a subclass of TextClass
 */
 
-#ifndef SHOES_TEXT_LINK_H
-#define SHOES_TEXT_LINK_H
+#ifndef SHOES_TEXT_LINK_TYPE_H
+#define SHOES_TEXT_LINK_TYPE_H
 
 /* extern variables necessary to communicate with other parts of Shoes */
 extern VALUE cShoes, cApp, cTypes, cCanvas, cWidget, cTextClass;
